Update streamlit_app.py
Put the list of selected fruits into a variable called fruits_selected. 

Then, use the fruits in our fruits_selected list to pull rows from the full data set (and assign that data to a variable called fruits_to_show). 

Finally, the app to use the data in fruits_to_show in the dataframe it displays on the page.